fix semaphore bug in present and acqure + add add_quark_application Cmake func

+++ Quark/RHI/Device.h
@@ -14,7 +14,7 @@ namespace quark::rhi {
 
     struct DeviceConfig 
     {
-        uint32_t framesInFlight = 2;
+        uint8_t framesInFlight = 2;
     };
 
     struct DeviceProperties
